rework form validator tests
