+++ src/Symfony/Component/Form/Tests/Extension/Validator/Constraints/FormValidatorTest.php
@@ -25,7 +25,6 @@ use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\SubmitButtonBuilder;
 use Symfony\Component\Translation\IdentityTranslator;
-use Symfony\Component\Validator\Constraints\Collection;
 use Symfony\Component\Validator\Constraints\GroupSequence;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\NotNull;
@@ -702,151 +701,6 @@ class FormValidatorTest extends ConstraintValidatorTestCase
 
         $this->assertCount(1, $context->getViolations());
         $this->assertSame($constraint, $context->getViolations()->get(0)->getConstraint());
-    }
-
-    public function testNonCompositeConstraintValidatedOnce()
-    {
-        $form = $this
-            ->getBuilder('form', null, [
-                'constraints' => [new NotBlank(['groups' => ['foo', 'bar']])],
-                'validation_groups' => ['foo', 'bar'],
-            ])
-            ->setCompound(false)
-            ->getForm();
-        $form->submit('');
-
-        $context = new ExecutionContext(Validation::createValidator(), $form, new IdentityTranslator());
-        $this->validator->initialize($context);
-        $this->validator->validate($form, new Form());
-
-        $this->assertCount(1, $context->getViolations());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[0]->getMessage());
-        $this->assertSame('data', $context->getViolations()[0]->getPropertyPath());
-    }
-
-    public function testCompositeConstraintValidatedInEachGroup()
-    {
-        $form = $this->getBuilder('form', null, [
-                'constraints' => [
-                    new Collection([
-                        'field1' => new NotBlank([
-                            'groups' => ['field1'],
-                        ]),
-                        'field2' => new NotBlank([
-                            'groups' => ['field2'],
-                        ]),
-                    ]),
-                ],
-                'validation_groups' => ['field1', 'field2'],
-            ])
-            ->setData([])
-            ->setCompound(true)
-            ->setDataMapper(new PropertyPathMapper())
-            ->getForm();
-        $form->add($this->getForm('field1'));
-        $form->add($this->getForm('field2'));
-        $form->submit([
-            'field1' => '',
-            'field2' => '',
-        ]);
-
-        $context = new ExecutionContext(Validation::createValidator(), $form, new IdentityTranslator());
-        $this->validator->initialize($context);
-        $this->validator->validate($form, new Form());
-
-        $this->assertCount(2, $context->getViolations());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[0]->getMessage());
-        $this->assertSame('data[field1]', $context->getViolations()[0]->getPropertyPath());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[1]->getMessage());
-        $this->assertSame('data[field2]', $context->getViolations()[1]->getPropertyPath());
-    }
-
-    public function testCompositeConstraintValidatedInSequence()
-    {
-        $form = $this->getCompoundForm([], [
-            'constraints' => [
-                new Collection([
-                    'field1' => new NotBlank([
-                        'groups' => ['field1'],
-                    ]),
-                    'field2' => new NotBlank([
-                        'groups' => ['field2'],
-                    ]),
-                ]),
-            ],
-            'validation_groups' => new GroupSequence(['field1', 'field2']),
-        ])
-            ->add($this->getForm('field1'))
-            ->add($this->getForm('field2'))
-        ;
-
-        $form->submit([
-            'field1' => '',
-            'field2' => '',
-        ]);
-
-        $context = new ExecutionContext(Validation::createValidator(), $form, new IdentityTranslator());
-        $this->validator->initialize($context);
-        $this->validator->validate($form, new Form());
-
-        $this->assertCount(1, $context->getViolations());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[0]->getMessage());
-        $this->assertSame('data[field1]', $context->getViolations()[0]->getPropertyPath());
-    }
-
-    public function testCascadeValidationToChildFormsUsingPropertyPaths()
-    {
-        $form = $this->getCompoundForm([], [
-            'validation_groups' => ['group1', 'group2'],
-        ])
-            ->add('field1', null, [
-                'constraints' => [new NotBlank(['groups' => 'group1'])],
-                'property_path' => '[foo]',
-            ])
-            ->add('field2', null, [
-                'constraints' => [new NotBlank(['groups' => 'group2'])],
-                'property_path' => '[bar]',
-            ])
-        ;
-
-        $form->submit([
-            'field1' => '',
-            'field2' => '',
-        ]);
-
-        $context = new ExecutionContext(Validation::createValidator(), $form, new IdentityTranslator());
-        $this->validator->initialize($context);
-        $this->validator->validate($form, new Form());
-
-        $this->assertCount(2, $context->getViolations());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[0]->getMessage());
-        $this->assertSame('children[field1].data', $context->getViolations()[0]->getPropertyPath());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[1]->getMessage());
-        $this->assertSame('children[field2].data', $context->getViolations()[1]->getPropertyPath());
-    }
-
-    public function testCascadeValidationToChildFormsUsingPropertyPathsValidatedInSequence()
-    {
-        $form = $this->getCompoundForm([], [
-            'validation_groups' => new GroupSequence(['group1', 'group2']),
-        ])
-            ->add('field1', null, [
-                'constraints' => [new NotBlank(['groups' => 'group1'])],
-                'property_path' => '[foo]',
-            ])
-        ;
-
-        $form->submit([
-            'field1' => '',
-        ]);
-
-        $context = new ExecutionContext(Validation::createValidator(), $form, new IdentityTranslator());
-        $this->validator->initialize($context);
-        $this->validator->validate($form, new Form());
-
-        $this->assertCount(1, $context->getViolations());
-        $this->assertSame('This value should not be blank.', $context->getViolations()[0]->getMessage());
-        $this->assertSame('children[field1].data', $context->getViolations()[0]->getPropertyPath());
     }
 
     protected function createValidator()
