+++ src/Symfony/Component/Form/Tests/Extension/Validator/ValidatorExtensionTest.php
@@ -12,23 +12,12 @@
 namespace Symfony\Component\Form\Tests\Extension\Validator;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\FormType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Validator\Constraints\Form as FormConstraint;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension;
 use Symfony\Component\Form\Extension\Validator\ValidatorTypeGuesser;
 use Symfony\Component\Form\Form;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormFactoryBuilder;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\GroupSequence;
-use Symfony\Component\Validator\Constraints\Length;
-use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Mapping\CascadingStrategy;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
-use Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory;
-use Symfony\Component\Validator\Mapping\Loader\StaticMethodLoader;
 use Symfony\Component\Validator\Mapping\TraversalStrategy;
 use Symfony\Component\Validator\Tests\Fixtures\FakeMetadataFactory;
 use Symfony\Component\Validator\Validation;
@@ -56,152 +45,5 @@ class ValidatorExtensionTest extends TestCase
         $this->assertSame(CascadingStrategy::NONE, $metadata->cascadingStrategy);
         $this->assertSame(TraversalStrategy::NONE, $metadata->traversalStrategy);
         $this->assertCount(0, $metadata->getPropertyMetadata('children'));
-    }
-
-    public function testDataConstraintsInvalidateFormEvenIfFieldIsNotSubmitted()
-    {
-        $form = $this->createForm(FooType::class);
-        $form->submit(['baz' => 'foobar'], false);
-
-        $this->assertTrue($form->isSubmitted());
-        $this->assertFalse($form->isValid());
-        $this->assertFalse($form->get('bar')->isSubmitted());
-        $this->assertCount(1, $form->get('bar')->getErrors());
-    }
-
-    public function testFieldConstraintsDoNotInvalidateFormIfFieldIsNotSubmitted()
-    {
-        $form = $this->createForm(FooType::class);
-        $form->submit(['bar' => 'foobar'], false);
-
-        $this->assertTrue($form->isSubmitted());
-        $this->assertTrue($form->isValid());
-    }
-
-    public function testFieldConstraintsInvalidateFormIfFieldIsSubmitted()
-    {
-        $form = $this->createForm(FooType::class);
-        $form->submit(['bar' => 'foobar', 'baz' => ''], false);
-
-        $this->assertTrue($form->isSubmitted());
-        $this->assertFalse($form->isValid());
-        $this->assertTrue($form->get('bar')->isSubmitted());
-        $this->assertTrue($form->get('bar')->isValid());
-        $this->assertTrue($form->get('baz')->isSubmitted());
-        $this->assertFalse($form->get('baz')->isValid());
-    }
-
-    public function testFieldsValidateInSequence()
-    {
-        $form = $this->createForm(FormType::class, null, [
-            'validation_groups' => new GroupSequence(['group1', 'group2']),
-        ])
-            ->add('foo', TextType::class, [
-                'constraints' => [new Length(['min' => 10, 'groups' => ['group1']])],
-            ])
-            ->add('bar', TextType::class, [
-                'constraints' => [new NotBlank(['groups' => ['group2']])],
-            ])
-        ;
-
-        $form->submit(['foo' => 'invalid', 'bar' => null]);
-
-        $errors = $form->getErrors(true);
-
-        $this->assertCount(1, $errors);
-        $this->assertInstanceOf(Length::class, $errors[0]->getCause()->getConstraint());
-    }
-
-    public function testFieldsValidateInSequenceWithNestedGroupsArray()
-    {
-        $form = $this->createForm(FormType::class, null, [
-            'validation_groups' => new GroupSequence([['group1', 'group2'], 'group3']),
-        ])
-            ->add('foo', TextType::class, [
-                'constraints' => [new Length(['min' => 10, 'groups' => ['group1']])],
-            ])
-            ->add('bar', TextType::class, [
-                'constraints' => [new Length(['min' => 10, 'groups' => ['group2']])],
-            ])
-            ->add('baz', TextType::class, [
-                'constraints' => [new NotBlank(['groups' => ['group3']])],
-            ])
-        ;
-
-        $form->submit(['foo' => 'invalid', 'bar' => 'invalid', 'baz' => null]);
-
-        $errors = $form->getErrors(true);
-
-        $this->assertCount(2, $errors);
-        $this->assertInstanceOf(Length::class, $errors[0]->getCause()->getConstraint());
-        $this->assertInstanceOf(Length::class, $errors[1]->getCause()->getConstraint());
-    }
-
-    public function testConstraintsInDifferentGroupsOnSingleField()
-    {
-        $form = $this->createForm(FormType::class, null, [
-            'validation_groups' => new GroupSequence(['group1', 'group2']),
-        ])
-            ->add('foo', TextType::class, [
-                'constraints' => [
-                    new NotBlank([
-                        'groups' => ['group1'],
-                    ]),
-                    new Length([
-                        'groups' => ['group2'],
-                        'max' => 3,
-                    ]),
-                ],
-            ]);
-        $form->submit([
-            'foo' => 'test@example.com',
-        ]);
-
-        $errors = $form->getErrors(true);
-
-        $this->assertFalse($form->isValid());
-        $this->assertCount(1, $errors);
-        $this->assertInstanceOf(Length::class, $errors[0]->getCause()->getConstraint());
-    }
-
-    private function createForm($type, $data = null, array $options = [])
-    {
-        $validator = Validation::createValidatorBuilder()
-            ->setMetadataFactory(new LazyLoadingMetadataFactory(new StaticMethodLoader()))
-            ->getValidator();
-        $formFactoryBuilder = new FormFactoryBuilder();
-        $formFactoryBuilder->addExtension(new ValidatorExtension($validator));
-        $formFactory = $formFactoryBuilder->getFormFactory();
-
-        return $formFactory->create($type, $data, $options);
-    }
-}
-
-class Foo
-{
-    public $bar;
-    public $baz;
-
-    public static function loadValidatorMetadata(ClassMetadata $metadata)
-    {
-        $metadata->addPropertyConstraint('bar', new NotBlank());
-    }
-}
-
-class FooType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('bar')
-            ->add('baz', null, [
-                'constraints' => [new NotBlank()],
-            ])
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefault('data_class', Foo::class);
     }
 }
